feat: 添加背景选中功能
