+++ src/app/chat/Background/index.tsx
@@ -1,3 +1,4 @@
+import { TRANSPARENT_ID, backgroundOptions } from '@/constants/background';
 import { useGlobalStore } from '@/store/global';
 import { useSettingStore } from '@/store/setting';
 
@@ -5,17 +6,18 @@ import { useStyles } from './style';
 
 const Background = () => {
   const { styles } = useStyles();
-  const backgroundImageUrl = useGlobalStore((s) => s.backgroundImageUrl);
+  const backgroundId = useGlobalStore((s) => s.backgroundId);
   const backgroundEffect = useSettingStore((s) => s.config.backgroundEffect);
+  const background = backgroundOptions.find((item) => item.id === backgroundId);
 
-  if (backgroundImageUrl) {
+  if (background && backgroundId !== TRANSPARENT_ID) {
     return (
       <div
         style={{
           position: 'absolute',
           left: 0,
           top: 0,
-          backgroundImage: `url(${backgroundImageUrl})`,
+          backgroundImage: `url(https://r2.vidol.chat/backgrounds/${encodeURIComponent(background.url)})`,
           backgroundSize: 'cover',
           backgroundPosition: 'center',
           width: '100%',
